# Conflicts:
#	MuseumSpring.iml
#	pom.xml
#	src/main/java/museum/controller/AuthorController.java
#	src/main/java/museum/controller/ExcursionController.java
#	src/main/java/museum/controller/ExhibitController.java
#	src/main/java/museum/controller/HallController.java
#	src/main/java/museum/controller/WorkerController.java
#	src/main/java/museum/service/ExhibitService.java
#	src/main/java/museum/service/impl/ExhibitServiceImpl.java
#	src/main/java/museum/service/impl/HallServiceImpl.java
#	src/main/java/museum/utils/ObjectMapperUtils.java
#	src/main/webapp/WEB-INF/static/errorMessage.jsp

+++ src/main/java/museum/service/impl/AuthorServiceImpl.java
@@ -7,8 +7,8 @@ import museum.dto.author.AuthorIdInitialsDto;
 import museum.dto.author.AuthorInitialsDto;
 import museum.entity.Author;
 import museum.exception.BadIdException;
+import museum.exception.EntityConstraintException;
 import museum.service.AuthorService;
-import museum.utils.ObjectMapperUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -26,13 +26,13 @@ import java.util.stream.Collectors;
 public class AuthorServiceImpl implements AuthorService {
 
   private AuthorDao dao;
-  private ObjectMapperUtils mapper;
 
   /** Method that save new author. */
   @Transactional
   @Override
   public void save(AuthorInitialsDto dto) {
-    dao.save(mapper.map(dto, Author.class));
+    dao.save(
+        Author.builder().firstName(dto.getFirstName()).secondName(dto.getSecondName()).build());
   }
 
   /**
@@ -58,7 +58,7 @@ public class AuthorServiceImpl implements AuthorService {
     if (author == null) {
       throw new BadIdException("Author has no row with id " + id);
     }
-    return mapper.map(author, AuthorFullDto.class);
+    return new AuthorFullDto(author);
   }
 
   /**
@@ -79,7 +79,13 @@ public class AuthorServiceImpl implements AuthorService {
   @Transactional
   @Override
   public void update(AuthorIdInitialsDto dto) throws BadIdException {
-    Author author = mapper.map(dto, Author.class);
+    Author author =
+        Author.builder()
+            .id(dto.getId())
+            .firstName(dto.getFirstName())
+            .secondName(dto.getSecondName())
+            .build();
+
     Author newAuthor = dao.update(author);
     if (newAuthor == null) {
       throw new BadIdException("Author has no row with id " + dto.getId());
@@ -90,10 +96,10 @@ public class AuthorServiceImpl implements AuthorService {
   @Transactional
   @Override
   public void deleteById(Long id) throws BadIdException {
-    Boolean isDeleted = dao.deleteById(id);
-
-    if (!isDeleted) {
-      throw new BadIdException("Author has no row with id " + id);
+    Author author = getOneById(id);
+    if (author.getExhibits().size() != 0) {
+      throw new EntityConstraintException("You can not delete this author because he is using.");
     }
+    dao.deleteById(id);
   }
 }
